Let Xom add an extra chaos brand to unrands (DracoOmega)
"This implements an item flag that lets unrand weapons be described as
'chaotic' and gives them the chaos brand (on top of whatever brand or
on-hit effects the unrand weapon might already have).

This is intended solely as a Xom perk for Xom's bazaars. Regular
rebranding can't work on unrands, and it felt like there could be some fun
to be had with Xom breaking the rules and giving you a touch more chaos on
some of the most interesting and exciting weapons."

[Committer's notes: This gives a 37.5% chance of seeing a chaotic-touched
normal unrand weapon distributed amongst the weapon types in a Xom bazaar,
and a 37.5% chance of seeing a regular unrand that was already risky or
chaotic enough without extra effects. Since chaos brand is actually a
relatively poor fit on weapons besides ranged or short blades or top-end
two-handers on a god that provides zero reliable emergency escape options,
the power already baked into unrands should help cover for it.

For the record, just to reiterate: absolutely do not provide this in any
contexts besides Sprints or Xom's bazaars. This is solely meant as a rare
treat for Xom worshippers in the base game.]