New Xom action: force speaking a slow Word of Recall
The easiest way to interact with players in a game about killing monsters
is to introduce more monsters, and while Xom's list of actions is very
full of summons already, there are yet still further contexts and twists
one can reasonably experiment with. This very rare bad Xom action, with
controls on minimum XL and maximum tension, should reasonably stand out on
a few fronts: it can be interrupted (and takes between 2x and 3x as long
as a normal word of recall to do so, so there's lots of space to react),
and it's pulling out actual monsters from the level and its vaults as
opposed to generating yet more monsters one could flee and wait out timers
for. Probably it's fine under Xom's continued flexibly unreliable nature?
Definitely not any worse than when a Zot trap randomly does it.

(Also, it allows lines like "The rat is forced to slowly start squeaking
a word of recall!", which is a great notion in and of itself.)