Remove 'attack previous target'  shout
'Attack previous target!' appears in the shout menu if you have a target, and they are alive in line of sight.

First, the 'Attack previous target!' shout can be issued even when not displayed on the menu; if you have a previous target alive but out of LOS it will command allies to target them. This technically leaks information, but is mostly just strange and buggy looking.

Second, what your previous target is is obtuse. Your target is changed when you perform any action which places a targetter, along with some other conditions. Notably, firing past a monster or casting apport will clear your previous target.

Given how low value ordering an attack on your previous target would be even if it were fully functional, it seems better to remove the option entirely.