Don't try to use the Bazaar wrapper outside of Bazaars (Cheibrodos)
The default-depth line for the entries was still applying to the bazaar
wrapper vault (for Xom versus not-Xom vault bazaars), which meant there
was a small chance for a bazaar-entry floor to try to place the wrapper
itself (and then fail due to the bazaar layouts all being only valid within
bazaars). Moving the default-depth line above it fixes this.