Pay HP/MP costs at the start of using an ability instead of afterward
Spells pay their costs before the code of a given spell executes, which
allows (for example) Vehumet/TSO's mp-on-kills passive to cover the cost of
the spell itself. But abilities only paid their costs *after* their code
was finished executing, meaning this was not possible.

It also resulted in oddness with Mark of Atrocity's scaling HP cost, where
the player would be charged the cost of the *next* level of atrocity while
only checking that they could afford to pay for the current level of
atrocity. (Fortunately it was not actually possible to die this way.)

I have rearranged code so that the payment happens before the ability is
used and is silently refunded if the ability fails or is cancelled (just
like spells currently do). Unfortunately, piety costs are much more
complicted, since gaining/losing piety can trigger messages and other
mechanical effects. So piety is solely paid after the ability code is
complete. I think this has much less of a noticable gameplay effect than
with HP/MP costs, however.

Finally, this also fixes a bug where the check for triggering Celebrant
after casting a spell with an HP cost happened at the wrong moment, and
could be triggered at instant speed merely by aiming a spell that *would*
result in the caster falling below 50% HP if they cast it - whether or not
they actually did so.