Tweak Dith weapon damage
Remove the damage reduction given to ranged weapons versus melee ones. I
was a little concerned originally that archer of Dith would be overly
strong compared to other options, but this seems to not be the case (and
many launchers are already slower than melee weapons, so they were
additonally disfavored).

Give quick blades a 25% reduction in shadow base damage compared to other
weapons. They should still be the weapon that benefits the most from shadow
mimic melee, but to a somewhat more moderate degree.