Slightly reduce Mark of Atrocity scaling cost, make finale auto-hit
Using ramping destruction with Mark of Atrocity is a heavy commitment that
costs a large amount of HP, and while I do like the risk/reward aspect of
this, it can feel unsatisfying if your big payoff misses. While I don't
think it was terribly likely to do so, let's try just making it not.