Tweak Crucible of Flesh danger level slightly
It's still a little hard to tell exactly where to balance this. On the one
hand, several players have repeatedly stated is was surprisingly easy to
handle. On the other, nearly 20% of players who have taken this mark have
ultimately died to the Crucible itself. Maybe this is slightly inflated by
how the same characters *without* Vessel of Slaughter may have died in
whatever situation caused them to use it instead.

At any rate, this is intended to be a fairly minor adjustment. The easiest
levels players can plausible enter the Crucible will have slightly weaker
demons, while at higher levels the Crucible has a slightly higher cap on
the number of demons it can produce at once (and how many will be there
when the player first arrives).