Buff Mark of Haemoclasm
The most underperforming of the Marks, haemoclasm is popular and
conceptually fun, but frequently failed to help the player very much - and
could in fact make things worse for them! The explosion chance was far too
low to rely on doing anything useful (and would commonly hit nothing even
when enemies did explode) but still generated normal explosion noise,
attracting more enemies to the player's position without even much
compensation.

This commit greatly reduces explosion noise (down to 5), lowers the chance
of enemies exploding with nothing nearby to hit, and greatly increases the
chance of enemies exploding when they *do* have something adjacent to hit
- rising moderately with how many things are nearby. I'm not 100% certain
that this is better than just a flatly higher chance at all times, but the
idea is to reduce purely-negative noise generation from blasts that cannot
be productive, while somewhat encouraging the player to wade into dense
piles of enemies to maximize explosion chance (and the possibility of
chain-reactions!) in crowds.

I've also slightly increased the blast damage caused by higher HP enemies.

We'll see how this feels and further adjustments will be made if called
for.