Fix Xom drained brain msgs.
They don't go through god_speaks(), so they need the substitutions and
substring handling to be done manually.

Maybe random body parts should be handled in the database instead of
the code, but that would require adding Lua hooks for all the functions
indicating which body parts the player does and doesn't have. For now,
this is simpler.

Closes #3966.