Fix a test suite error (unhappily)
It was being used to perform mutation tests without a core part of mutation
data being initialized, and only appeared to work up to this point due to
what I consider essentially a coincidence. It made mutation-related
functions effectively consider *all* mutations to have the data of the
first mutation listed in mutation-data.h, which until 02fa12cfa58bf5b15c769780ce23e0ab14f10ce6
was Tough Skin (which happened to have 3 levels - the only reason that a
test didn't fail before now).

The test harness here is skipping a massive amount of standard
initialization code for Crawl, seemingly deliberatetly. I understand the
desire for more isolated tests, but this strikes me as deeply concerning,
as it's not at all clear when necessary parts in this complex system may or
may not be initialized in ways that cause the test to produce incorrect
values. This time, a test failed that should have succeeded, but it seems
entirely possible to me that it could also cause tests to succeed that
ought to fail.

Basically: I don't really like this, or my equally ad hoc fix here. This
will silence build failures for the moment, but it feels like a bad
long-term solution to any of this.