Make dryads cautious
Summon Forest is already a strong spell, but the degree to which the dryad
(key to the spell actually doing anything at all) will suicidally rush
enemies for no benefit can sometimes be fiddly or dissatisfying. Having to
protect them with your own body can be interesting, but the open layout
often makes that impossible when they're so insistant on moving past you
to hit stuff for irrelevant damage, and it feels wrong that it's helpful to
order them *not* to attack stuff, which you can only do if you're not using
other allies at the same time.

Now they should be a little less overeager to close in on enemies.