Remove a digging exploit with formicids of Jiyva
While wands of digging would simply not work on temporary terrain prior to
the previous commit, formicid digging *did* work. And it was possible to
use Oozemancy to make metal walls slimy and then simply walk through them.

While it would be, in theory, possible to apply the same policy as the
previous commit - let the player walk through walls whose underlying
feature could also be dug - it has awkwardness with how the player moves in
the process. We can't just revert to metal and keep you in the same place;
the player may have forgotten which walls were metal and which weren't!

After a little discussion, it seems fine to dodge the question by just not
letting you dig through acidic slime with your mandibles at all.
Technically a small nerf to formicids in Slime, but I doubt it ever really
mattered there anyway. (I am normally adverse to one-off exceptions like
this, but people seem to think this one is intuitive enough and it neatly
solves the issue.)