Show some status icons above the player (and add customizable options)
Certain negative statuses are very important, but can be easy for a player
to overlook when they are applied randomly or unexpectedly (eg: Slow from
*Slow triggering, or from draconian cold-bloodedness). We already have
standard icons for many of these things, so it seems possible to indicate
them directly above the player, where it may be a lot more obvious than
the sidebar.

I have added support for displaying a number of statuses there (and it is
now straightforward to add more). Since there are many statuses that
*could* be indicated but which are often not important to do so (either
because they're low impact, beneficial, or already have their own default
force_more), I've added a new option: tile_player_status_icons. This allows
a player to decide which of the implemented statuses they want to show this
way (if any).

Currently there is support for displaying slow, fragile, sentinel's mark,
will/2, haste, weak, corrosion, might, brilliance, -move, constriction and
petrifying. By default, it only displays slow, fragile, constricted, and
1/2 will. This is somewhat arbitrary and I am open to input for different
defaults.

I tried to make the string used in the options to enable a status be the
same as it appears on the sidebar (though it bothers me a bit that it's
harder than I'd expect to efficiently query this directly - duration-data.h
currently cannot be included anywhere without manually including a bunch of
things before it, which is awkward....)

-Potion and -Scroll are other statuses that immediately jump out to me as
worth showing in this manner (since they can be critical, and also easy to
miss activating) but no icons exist for them yet. (Also, poison is still
handled by an older method and cannot be edited with this option yet.)