Make it more obvious where the Crucible exit is
Properly update the minimap when the portal opens (I thought this happened
already, but apparently not) and also make a ring around it change to a
different type of floor. Possibly a higher-visibility floor could be
chosen or (ideally) a fancier exit tile made at some future point, but this
should still improve visibility a little, at least.