Properly render clouds on top of items in local tiles
The comments suggest this was always intended, but didn't actually work due
to how items are packed in a different buffer than monsters (and one which
appears to always be rendered after clouds).

Now it applies a similar technique used when items are to be drawn
submerged in water, and moves cloud-engulfed items to a lower buffer.

(Webtiles uses different rendering code and did not have this bug.)