Increase Cigotuvi's Putrefaction draining cost
The latest iteration on corpse rot seems to have been well-received, but is
probably a bit strong for where it is. I considered moving it to level 5,
but the original intent of "Strong for a level 4 spell, but with a
meaningful draining cost" is something I want to try leaning a bit more
into first, and the current iteration virtually removes the cost if
spellpower is high enough.

New draining cost is 75->35 at max power (versus 55->5 before this commit)