Properly make orc wizard apostles Cautious
Astoundingly, despite this flag having been implemented *for* them in the
first place, it appears to have never actually been implemented in trunk
at any point. I am really not sure how this came to happen.

Hopefully this makes this feel a little less suicidal to have as allies,
as it was originally intended to.