Slightly relax Eringya's Toxic Bog's placement restrictions
By allowing bog to be created on tiles with a maximum of *2* adjacent walls
instead of just one. Bog is a spell with a lot of limitations, and while I
think some of those limitations are interesting, it could probably stand to
be just a *bit* more permissive. Let's give it a try.