Tweak some Xom messages
The Xom pseudo-miscasts that entail "Something invisible lands on the
floor beside you!" and the like unfortunately collide with the warning
colours in the default messages.txt for invisible monsters using
something as broad as "Something .* you". This commit replaces those
messages with new ones that won't collide with such and are still
appropriately confusing for Xom.