Brief quick vault passover
Some living trees replaced with dead ones, a late D glowing orange brain
placement nerf, and some decorative altar adjustments + fix-ups.

The last of these adds a small chance for the tiny altar square used in
a large amount of games to get a highly variable floor tiling arrangement
around it, to add a little more interest to the open-altar squares being
used constantly throughout such a large part of the game. It also comes
with a few tweaks to _pick_an_altar() for a few branches to catch up to
some later flavour options; as long as we've got such a minor flavour
function already around, we might as well continue to update it.