New option: food_snacking_frequency (default 40%)
A relatively quick and simple setting for caches of fruit and meat, which
allows setting the percent chance of displaying their corresponding flavour
message. Should be relatively harmless, but also reduce the current
awkwardness of people who want to see the messages dancing back and forth
on top of the piles (as long as they read the options guide, alas).

On that same note, since people seem to be relatively receptive of the
flavour messages in the first place, the odds have been increased (from
~33.3% to 40%).