Add a force_more to sighting the_gamble (Riverfiend)
While autoexplore does stop for the immediate sightings of both an altar
and a transporter, it is relatively plausible to completely miss that it is
a limited time offer amongst the messages of sighting the enemies and
rewards in either side. Since this has completely different behaviour than
normal transporter vaults in being temporary (and is thus better off for
it), it might as well get a force_more the way temporary portals do.