Change Dazzling Flash schools (Conj/Hexes -> Hexes/Fire)
A remnant of back when Dazzling Spray actually *was* a damaging spell,
conjurations has been an extremely weird fit on the completely non-damaging
Dazzling Flash spell for a while. It also meant that Vehumet could gift it
to players, despite it not being appropriate to the god at all.

The choice of Fire as the new secondary school is somewhat arbitrary, but
it felt like the closest thematic fit for 'makes a lot of light, very
abruptly'. (I don't think it will affect players much, as one never really
trained conjurations as a hexer, nor picked up flash as a conjurer.)