Adjust Dazzling Flash tracer for monsters, increase duration vs. player
Monsters previously considered already-blind enemies as useful targets for
this spell. And while the duration *can* stack, it's generally not very
useful to keep extending blindness on an already blind player (and, in
fact, the moon troll's cast rate was high enough that they would spend
fully 25% of their turns casting it, regardless of it accomplishing
anything).

In some mild compensation, increase the baseline duration when applied to
the player. This does affect some other methods of blinding the player, but
in general it is a weaker effect on players and a longer duration makes
sense to me. (This also doesn't matter a lot given the rarity of blind
effects in the game that will hit the player. This is slated to be changed
somewhat in 0.33 and duration will be reviewed again at that time.)