Allow divine heal-on-kill to work on monsters which are 'objects'
It already seperately excluded summons and things which don't grant the
player XP, but for some reason it *additionally* vetoed monsters which
were objects. Monsters that counted as objects but also gave XP *mostly*
meant 'dancing weapons', but it was still strange that Makh/Veh/TSO healing
didn't count them.

This bit of code seems ancient, and if there was ever a good reason for it,
I doubt there is anymore.