Make a few more spells work with Aphotic Marionette
This implements Marionette casting of Battlesphere, OTR, Polar Vortex, and
Malign Gateway (mostly through calling the player versions of these spells
via overrides).

It also fixes Berserker Rage being inconsistent about when it could be used
this way, as well as Invisibility not producing a message when gained this
way.