Give RevMPSaver an additional effect, rename to SpellMotor
In theory, flat spell cost reduction is a powerful effect, but capitalizing
on this gizmo property is difficult. To benefit from it, you need to engage
in weapon combat first, but many of the spells which gain the most from
being able to repeatedly cast with cost reduction tend to be attack spells
you'd preferred to have cast from a distance. Its design seems to want a
character who is strong enough to comfortably sit in melee, but who has
relevant blasty spells they want to cast a lot *after* that point, which
is a hard sell on a species whose biggest strength comes from leaning into
their ability to do enormous weapon damage while sacrificing defense.
Moreover, the cost reduction is also short-lived, since Rev will wear off
while the player is casting.

In the hope of moderately addressing both of these awkwardnesses, I am
adding a second effect to RevMPSaver and renaming it to SpellMotor. In
addition to its current effect, it also causes the player to launch a
melee attack at a random enemy in range whenever they cast a spell. This
gives a direct boost to the hybrid blaster archetype, but also helps the
player *sustain* the cost reduction by maintaining Rev if there are nearby
enemies to hit while casting. It also allows the player to *build* Rev by
spellcasting (but only while directly engaging in combat with those
spells).

If your weapon attack would somehow have a delay *longer* than it takes to
cast a spell, there is a proportional chance to perform no attack (so no
being cheeky with giant heavy weapons).

While this arguably makes this gizmo property more powerful than its peers,
I think the archetype it directly supports is also weaker and less
straightforwardly popular on coglins such that I think this is okay. (And
it would be nice if all gizmo properties felt reasonably worth taking; I
think RevMPSaver is currently a bit of a dud.)