Fix Unleash Destruction being completely silent
Not sure *exactly* how much noise to give it, but 'more than none'. In
theory, I could see it argued that lightning-based destruction should be
louder than other types, and this wouldn't actually be hard to implement,
but I'm unconvinced that it's meaningfully better to sometimes be much
louder than other times at random.