Make uniques a little more chatty
There's been a bunch of effort in recent years to characterize uniques
better, and give them more personality in their dialogue. It seems fair
that they get to chat a little more than other monsters as a consequence,
so that players have a better chance to see it before killing them.

(Donald and Yuif had greatly increased odds to talk compared to other
uniques, and I have lowered their bonus to keep their net rate of
conversation about the same as it was before now.)