Fix enemies being mislead by Shadowslip after being charmed (Monkooky)
If an enemy was distracted by your shadow, they would continue to fixate
on it even after being charmed. On the one hand, this makes a certain
sense. They think your shadow *is* you, so why should they listen to orders
from that othere chump over there? And it *is* pretty funny.

But probably it should just break the mislead effect, so now it does.

This fixes #3970