Don't let !/@ work when aiming smite-targeted spells
These keys allow force-casting a spell not just beyond its range, but at
places where the spell's targeter:::valid_aim() returns false - essentially
ignoring any target validity checks that are done there. While this serves
some purpose for beam spells, by allowing the player to select beam paths
that are not possible within the spell's range, (and the beam's range and
physical properties will be respected, regardless), I don't think it
serves any purpose for spells which target a specific location. Forcing a
spell to be cast on a location the game itself says is not valid is just
asking for undefined behavior. And boy, there's quite a bit of it at the
moment. This may not be comprehensive, but at the moment at the very least:

-Grave Claw, Mercury Vapours, Momentum Strike, Cigotuvi's Putrefaction,
 and Sculpt Simulacrum can all be cast at full LoS range.
-Magnavolt and Putrefaction can trigger a crash-on-demand by aiming at a
 tile without a monster.
-Putrefaction can be functionally used on uninjured monsters and non-living
 monsters.

Many older spells in the game do not exhibit the same range bugs because
each spell's own code checks manually that the target they have been given
is valid and in range - with even the most straightforward of range checks
being essentially duplicated countless times in different places. Some of
these spells predate the targeting API existing at all, and some were
bandaided later on, but I'm inclined to think that where a spell defines a
hitfunction for validity, that this function should be authoritative and
the spell should not have to duplicate its checks (this sort of problem has
come up many times in past and been fixed in a one-off manner on many of
this times.)

Doing so would allow cutting out a bunch of now-redundant code from many
spells, though I am going to pass on doing so during feature freeze, since
it seems at risk of introducing new bugs. I'd like to make it a 0.33 thing,
though.