Expire boulders and ballisto spores when changing floors
At some point in the not-so-distant future I will actually unify 'temporary
monster' checks that don't depend specifically on abjurability. But in the
meantime, don't let the player drag boulders between floors with them.

(Amusingly, spores would effectively only follow you only if there was a
hostile monster in sight when you took the stairs, and otherwise stay on
their original floor.)