Some Gavotte targeter improvements
If a monster that will get wall slammed overlaps with the player's transit
path, it will still be hilighted to indicate it will take damage.

The targeter now accounts for seen terrain and monsters that are outside
the player's current line of sight, instead of effectivey treating
everything the player cannot currently see as empty floor. This allows more
intuitive and accurate hilighting of which monsters will take damage from
a given movement (since it otherwise may not be obvious why a given monster
isn't hilighted when the *player* knows there's a wall in the way already.)

This does mean the tracer will sometimes claim a monster will take damage
when it won't (ie: because something dug a wall out of line of sight or
destroyed a plant it was expected to hit) but this should be much, much
more rare than the previous situation of ignoring map memory of unseen
walls entirely.