Fix an issue with Mark of the Celebrant and Reflection
Since Celebrant's cooldown was set *after* the volley of blood arrows
fired, if the player took damage as a result of it, this damage could
trigger another barrage. (Possibly this could also happen with Pain Mirror)

But while this is a bug in its own right, the crash mentioned in #3972
doesn't seem to me like it could be directly caused by getting multiple
bloodrites. I think reusing the beam structure may have resulted in
improper values being set if the beam was reflected in the process. So
instead, copy the beam structure anew with each shot to avoid this.

This should fix #3972