Stop Coglins from destroying cursed items without a prompt in local tiles
Clicking on a weapon in the inventory panel in local tiles attempts to
wield that weapon. For coglins in particular, it bypasses the normal
prompt for 'which hand do you want to put that weapon in?' - with left
click going into the main hand and ctrl+left click going into the offhand.

However, the logic in can_wield() regarding curses only checked that the
player had *a* hand free to wield a given weapon - not necessarily the hand
they were about to put something in! So if a player had one cursed weapon
and their other hand was free, using click-to-equip for the hand with the
cursed weapon would see that it was possible to equip a new weapon, and
then immediately unwield and destroy the cursed item to replace it with the
one just clicked on.

Now there is specific logic to check for the corresponding weapon being
cursed first.