Don't (confusingly) leak invisible monsters with Piledriver (Acrobat)
While Piledriver already could never target and move a monster the player
could not see, it was possible for an invisible monster to be in the way of
moving something the player *could* see, which resulted in a confusing
message about the spell being unusable in situations where it looked like
it should be.

Now the targeter will display paths based on what the player can see, but
if an invisible monster is in the way of the selected path, the spell will
fizzle without effect when cast instead. This is a little unfortunate, but
also we can't really move something that has nowhere to move. (If *a*
viable path remains unblocked, the spell will take that one instead)