Update removed_spells list
And add version tags around a few spells that were effectively removed, but
never got them.

I wonder if, instead of using the AXED_SPELL macro in spell-data.h, dummy
data should automatically be provided for any spell in the removed_spell
list? Since currently there are effectively multiple places to say that
the spell was removed, with different consequences for not being listed
there.