Show some fprops in the local tiles tile tooltip
It wasn't easily possible to see what tiles were marked no_tele_into in
game, and this is helpful for diagnostics sometimes.

(I didn't bother showing fprops here which are easy to verify visually,
such as Sanctuary. It would be easy enough to expand this list in future,
though, if there's any need.)