Allow no_tele_into tags (and some others) to actually work for subvaults
Map tags which applied fprops (such as no_tele_into, no_tide, no_jiyva)
did nothing at all when attached to a subvault. Most of these were not
used this way, as far as I can tell, but no_tele_into *was* used several
times - notably in the S-branch preview chambers for
grunt_ashenzari_visionary - meaning that it was possible for the player to
teleport (or take a hatch) into a place from which there was no escape and
also many wildly out of depth monsters.

Now, the same set of fprop tags that work with parent vaults should also be
respected by subvaults.

(It's worth nothing that kprops applied to subvaulted glyphs by the parent
vault are *also* ignored when a subvault is merged, so there was really no
good way to do this sort of thing prior to this commit, despite what one
might expect.)

This fixes #3979