Fix Xom glove slot msgs for certain randarts.
The Mad Mage's Maulers and Delatra's gloves don't start with "pair of ",
but make Xom glove slot messages put that at the beginning of their
names anyway, so that all glove names come out as singular and the
message grammar fits.

There should be a better way to handle this, but this will work for now.