fix: Don't mark certain spells as self enchantment
A number spells are marked as self enchantments despite not setting any
kind of duration or enchantment on the caster itself. These are mostly
demon-related summon spells, a few conjurations that create monsters
(e.g. Marshlight and Conjure Ball Lightning), as well as Phantom Mirror
and Spawn Tentacles. For the non-summoning spells, having this flag
leads to them not being considered as ranged spells. For Spawn
Tentacles, there's even an explicit check to override this! This commit
removes the self enchantment spell flag on the aforementioned spells. It
also removes checks in the monster casting code for some of these spells
that were only needed because of this flag.

One subtle aspect affected by this commit is a monster's second spell
cast attempt. If a monster rolls a targeted spell for its action and is
unable to find a target, a second casting attempt is made, and 50% of
the time this second check only considers self enchantment spells. So
this change effectively reduces the frequency monsters would cast these
spells if they also have a spell that can somehow fail to target.
However this frequency reduction is very small in practice for the
monsters affected by this change. For example, I have tested orc
sorcerers who are in LOS but unable to target the player with bolt of
drain/paralyze, and they still readily cast Summon Demon. Monsters that
don't have targeted spells in addition to these spells will see no
change at all.

Not including summons nor spells that make conjured monsters as ranged
attacks doesn't make a lot of sense and should be revisited in a future
commit. In terms of player safety, many monster summoners create summons
with "proper" ranged attacks or superior terrain traversal compared to
the summoner. And conjured monsters certainly pose an acute threat to
the player at range. I want to make extra sure I understand all the
implications of designating summon spells as ranged attacks for
monsters, although I'm fairly sure there are no issues. For now, the
only change in this regard is removing an explicit exclusion for Conjure
Ball Lightning as a ranged spell. Lom Lobon is the only monster with
this spell and has Glaciate, another ranged spell, so this has no
in-game effect.

Finally, Repel Missiles and Berserker Rage are clearly self enchantments
and so are given the self enchantment flag by this commit. For Berserker
Rage, this means that bear monsters will no longer be flagged as ranged
monsters.

(cherry picked from commit fa02dad217768b62d142e339f2926be0b0653075)