Added a batch of 44 vault decor sprites (#3992)
"including baked goods caches, flower pots, banners, fountains, and a mop"

[Committer's notes: Pretty great tiles that'll help spruce up a large
number of vaults and portals and branches. Squashed the three commits
together, put the wall tiles into their own folder, made one additional
variation of a flower garden tile going into UNUSED for future plans, and
updated some comments + moved it to the right position in the text file.
Following commits will make use of a number of the more straightforwardly
usable tiles. Not entirely sure how to handle some of the others, but the
overall set can be useful for other devs in the future.

Closes #3992.]