Fix a crash when trying to use quivvered Shadowslip with no shadow
Even if the ability was unusable because of no active shadow, the codepath
for using it via quiver would still construct a targeter first, and this
targeter (used to show what monsters will be distracted) assumed that it
could not be called without an active shadow. This turns out to not be the
case.

(The targeter is never shown to the player, either way.)