Fix Shadowslip having a misleading duration when used repeatedly (Acrobat)
Using Shadowslip while it was already active would *set* the duration of
the misdirection effect on all enemies to a rolled duration, while the
visual change to your shadow was *added* to the duration of its curent
transformation. This meant that if you used it again before it wore off,
the effect would expire long before it 'looked' like your shadow had
returned to normal.

Now the shadow visual transformation should better match the duration of
the mechanical effect.