Display which monster is engulfing the player (when examining them)
In a similar fashion to when a monster is constricting you, indicate which
monster is engulfing you when you are being engulfed (to help with
ambiguities when there are several possible monsters nearby).

This doesn't indicate the source of monster/monster engulfment, but that
seems of much more marginal value. (It is much less straightforward to do
this, since monsters themselves don't track who or even *if* they are
currently engulfing anything)