Make recruitable apostles disappear when you change floor
Previously, if you were off-level when the recruit timer expired, the
apostle would remain wandering on that floor forever. Instea, end the timer
immediately if you leave the floor.

I'm a *tiny* bit uncertain about whether this could catch a player
off-guard by ending the window prematurely, but it's relatively consistent
with other uses of good-neutral orcs and there's probably not many
situations where a player needs to leave a floor immediately, and still
plans to come back and recruit afterward. (And this is simpler to
implement by far).