Fix a possible crash with Xom fake shatter
It tried to modify the quantity of the large rocks created *after* they
were placed on terrain, but if the location selected to place them was
lava, the rock would already be destroyed by the time the code attempted to
modify it.

It would be easy enough just to not try to place rocks in lava at all, but
the sizzling splashes caused by them seems flavorfully appropriate enough
to me.