Add Grave Claw to the list of monster smitey spells
...that is used to determine whether the caster will push past allies.

As per the comment above the list, these probably should really be spell
flags (or maybe a flag on the monster itself instead??) There are quite a
few things which would theoretically qualify to be on this list, but I'm
not filling it out right now as this would cause monster behavior changes
that may or may not be positive and shouldn't be done so close to release.

Will aim to do something with it in 0.33, though.