fix: "Ran out of altars for temple" error (thirdmesn, Vajrapani) (#4005)
A few temple vaults had a temple_altars_X tag where X was lower than
the number of altars in the vault. This caused the "Ran out of altars
for temple" error during the creation of a new character/dungeon.

This commit fixes the issue by increasing the tag values to match the
number of altars in the vault. This change results in temple vaults
with 22 and 26 altars, instead of the previous max of 21.

Note: Does not update temple vaults with a variable temple count due
to substitution removing the superfluous altars in the vault.

Resolves #2810
Resolves #3846

[Committer's note: actually tweaked the initialization temple function to
allow the 22-altar temples to be picked. This _does_ slightly raise the
average number of gods compared to 1659c76, but it's probably fine now
that more gods are attractive to more characters in 0.32. Also, left a few
more comments around.

Closes #4005.]