Don't trigger chaotic artifact weapons against dead monsters
This could happen if specifically the base *brand* of the weapon killed a
monster, and even then didn't usually cause a visible problem. With chaotic
Rift and Punk *specifically* (among all chaotic unrands that can currently
appear in the Xom bazaar), this resulted in a crash.

Unlike other sources of brand damage, Punk's acid and distortion banishment
fully reset a monster when they happen, meaning the defender's type is
MONS_NO_MONSTER by the time the secon apply_brand_damage() happens. And
chaos brand specifically tries to look at a monster's resists to determine
what brands it can roll. Against a normal dead monster, this happens
normally (and then it elects not to apply the brand at all), but against a
reset monster, this causes an assert when it looks for its resists.

I think a more proper solution involves refactoring the chaos brand check,
but I'd like to cherry-pick a bugfix in 0.32, so I'm going with the
simpler solution with less possible side-effects for right now.