Further adjust Cheibriados wrath and haste response
Cheibriados wrath uses tension very strangely with a both fine-grained yet
randomized fashion. Since tension is useful only as an extremely loose
assessment and awkward to ever calculate versus given threats, I'm going
to heavily simplify the check, so it will hopefully reduce the need to
keep an close eye on this wrath whenever adjusting tension in the future.

The five different tiers of mixing together effects across Cheibriados's
current wrath have been reworked and combined into three instead.

  * If there's almost any notable tension:
    * current hp >= 75% of your max hp: inflict sleep.
    * Otherwise: remove haste plus inflict slow.
  * Otherwise: drain stats plus make an extremely loud noise.

This is still dynamic in its own straightforward fashion, hopefully:
bringing in more enemies while exploring, threatening the player without
just using direct damage while in combat in a fashion that can
technically be survived, and statdraining the player heavily if they try
to wait it out in a closet. To disincentivize the last of these, stat
drain has been further pushed up in severity.

(Cheibriados wrath has been repeatedly heavily changed, perhaps because
the time theme is one of the loosest in available themes for malevolent
effects- wrath once contained miscasts, other god wraths, random item
curses, and so on and so forth. God wrath doesn't inherently need the most
variety as long as it's actually reasonably threatening in a distinct
fashion, though, as can be seen with Hep and Ash and Ru, so it should be
reasonably fine in this very compressed state.)

Also in this commit: Cheibriados now prevents potions of haste from ever
actually affecting the player, rather than forgiving them only if they
drink it unknown and throwing them into penance otherwise. This is mostly
meant to makes matters more consistent with Chei's prevention of haste
from chaos and berserk, as well as with Okawaru's prevention of using ally
summoning scrolls. In the future, it also helps with probably just making
all penance just involve removing god passive benefits and invocations as
the good gods and Beogh do, as piety loss to a sufficient degree and said
restrictions can be plenty enough downsides, and it'd make god penance
fairly more straightforwardly parseable for all gods.